feat: cache, tests

+++ internal/model/payment.go
@@ -6,7 +6,7 @@ type Payment struct {
 	Currency     string `json:"currency" db:"currency" validate:"required"`
 	Provider     string `json:"provider" db:"provider" validate:"required"`
 	Amount       int    `json:"amount" db:"amount"`
-	PaymentDt    int    `json:"payment_dt" db:"payment_dt" validate:"required"`
+	PaymentDt    int64  `json:"payment_dt" db:"payment_dt" validate:"required"`
 	Bank         string `json:"bank" db:"bank" validate:"required"`
 	DeliveryCost int    `json:"delivery_cost" db:"delivery_cost" validate:"required"`
 	GoodsTotal   int    `json:"goods_total" db:"goods_total"`
